+++ src/app/_shared/services/api-data/api-data-schedules.service.ts
@@ -20,11 +20,11 @@ export class ApiDataSchedulesService {
     .map((response: Response) => response.json());
   }
 
-  updateProjectSchedule(schedule: any, revisionNotes: string): Observable<any> {
+  updateProjectScheduleXML(schedule: any, revisionNotes: string): Observable<any> {
     const headers = new Headers({'Content-Type': 'application/json', 'X-Token': this.cacheService.token.signedToken});
     const options = new RequestOptions({ headers: headers });
 
-    return this.http.post(`api/updateProjectSchedule/${revisionNotes}`, JSON.stringify(schedule), options)
+    return this.http.post(`api/updateProjectScheduleXML/${revisionNotes}`, JSON.stringify(schedule), options)
     .timeout(this.cacheService.apiDataTimeout)
     .map((response: Response) => response.json());
   }
@@ -38,22 +38,62 @@ export class ApiDataSchedulesService {
     .map((response: Response) => response.json());
   }
 
-  updatePartSchedule(schedule: any, revisionNotes: string, userID: number): Observable<any> {
+  updatePartScheduleXML(schedule: any, revisionNotes: string): Observable<any> {
     const headers = new Headers({'Content-Type': 'application/json', 'X-Token': this.cacheService.token.signedToken});
     const options = new RequestOptions({ headers: headers });
 
-    return this.http.post(`api/updatePartSchedule/${revisionNotes}`, JSON.stringify(schedule), options)
+    return this.http.post(`api/updatePartScheduleXML/${revisionNotes}`, JSON.stringify(schedule), options)
     .timeout(this.cacheService.apiDataTimeout)
     .map((response: Response) => response.json());
   }
 
-  destroySchedule(scheduleID: number): Observable<any> {
+  destroyScheduleSP(scheduleID: number): Observable<any> {
     const headers = new Headers({'Content-Type': 'application/json', 'X-Token': this.cacheService.token.signedToken});
     const options = new RequestOptions({ headers: headers });
 
-    return this.http.get(`api/destroySchedule/${scheduleID}`, options)
+    return this.http.get(`api/destroyScheduleSP/${scheduleID}`, options)
     .timeout(this.cacheService.apiDataTimeout)
     .map((response: Response) => response.json());
+  }
+
+  insertSchedule(scheduleData: any, userID: number) {
+    const headers = new Headers({'Content-Type': 'application/json'});
+    const options = new RequestOptions({ headers: headers });
+    return this.http.post(`/api/insertSchedule/${userID}`, JSON.stringify(scheduleData), options)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
+  updateSchedule(scheduleData: any, userID: number) {
+    const headers = new Headers({'Content-Type': 'application/json'});
+    const options = new RequestOptions({ headers: headers });
+    return this.http.post(`/api/updateSchedule/${userID}`, JSON.stringify(scheduleData), options)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
+  destroySchedule(scheduleData: any, userID: number) {
+    const headers = new Headers({'Content-Type': 'application/json'});
+    const options = new RequestOptions({ headers: headers });
+    return this.http.post(`/api/destroySchedule/${userID}`, JSON.stringify(scheduleData), options)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
+  insertScheduleDetailBulk(scheduleData: any, userID: number) {
+    const headers = new Headers({'Content-Type': 'application/json'});
+    const options = new RequestOptions({ headers: headers });
+    return this.http.post(`/api/insertScheduleDetailBulk/${userID}`, JSON.stringify(scheduleData), options)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
+  }
+
+  updateScheduleDetailBulk(scheduleData: any, userID: number, scheduleID: number) {
+    const headers = new Headers({'Content-Type': 'application/json'});
+    const options = new RequestOptions({ headers: headers });
+    return this.http.post(`/api/updateScheduleDetailBulk/${userID}/${scheduleID}`, JSON.stringify(scheduleData), options)
+      .timeout(this.cacheService.apiDataTimeout)
+      .map((response: Response) => response.json());
   }
 
 }